+++ client/components/index.js
@@ -10,4 +10,3 @@ export {Login, Signup} from './auth-form'
 export {default as Home} from './Home'
 export {default as SearchBar} from './SearchBar'
 export {default as Graph} from './Graph'
-export {default as SingleGraph} from './SingleGraph'
